Structure the frontend testing framework
