+++ packages/ballerina-visualizer/package.json
@@ -100,6 +100,7 @@
     "webpack-cli": "5.1.4",
     "webpack-dev-server": "5.2.4",
     "@types/jest": "29.5.14",
+    "@types/node": "22.15.21",
     "jest": "29.7.0",
     "ts-jest": "29.3.4",
     "@testing-library/react": "16.3.0",
