+++ packages/ballerina-visualizer/tsconfig.json
@@ -39,10 +39,6 @@
   "include": ["src"],
   "exclude": [
     "node_modules",
-    "build",
-    "src/**/*.test.ts",
-    "src/**/*.test.tsx",
-    "src/**/*.spec.ts",
-    "src/**/*.spec.tsx"
+    "build"
   ]
 }
